Extracted TypeRegistryReverseMap and added accessors
Authored-by: Donal Evans <doevans@pivotal.io>
