+++ geode-core/src/main/java/org/apache/geode/pdx/internal/TypeRegistry.java
@@ -562,4 +562,12 @@ public class TypeRegistry {
   Map<Class<?>, PdxType> getLocalTypeIds() {
     return localTypeIds;
   }
+
+  Map<Integer, EnumInfo> getIdToEnum() {
+    return idToEnum;
+  }
+
+  Map<EnumInfo, Integer> getEnumInfoToId() {
+    return enumInfoToId;
+  }
 }
