+++ geode-core/src/test/java/org/apache/geode/pdx/internal/PeerTypeRegistrationReverseMapTest.java
@@ -23,39 +23,7 @@ import org.junit.Test;
 
 import org.apache.geode.cache.Region;
 
-public class PeerTypeRegistrationReverseMapTest {
-
-  @Test
-  public void saveCorrectlyAddsOnlyToReverseMaps() {
-    PeerTypeRegistrationReverseMap map = new PeerTypeRegistrationReverseMap();
-    assertThat(map.typeToIdSize()).isEqualTo(0);
-    assertThat(map.enumToIdSize()).isEqualTo(0);
-    assertThat(map.pendingTypeToIdSize()).isEqualTo(0);
-    assertThat(map.pendingEnumToIdSize()).isEqualTo(0);
-
-    addPdxTypeToMap(map);
-
-    assertThat(map.typeToIdSize()).isEqualTo(1);
-    assertThat(map.enumToIdSize()).isEqualTo(0);
-    assertThat(map.pendingTypeToIdSize()).isEqualTo(0);
-    assertThat(map.pendingEnumToIdSize()).isEqualTo(0);
-
-    addEnumInfoToMap(map);
-
-    assertThat(map.typeToIdSize()).isEqualTo(1);
-    assertThat(map.enumToIdSize()).isEqualTo(1);
-    assertThat(map.pendingTypeToIdSize()).isEqualTo(0);
-    assertThat(map.pendingEnumToIdSize()).isEqualTo(0);
-
-    Object fakeKey = mock(Object.class);
-    Object fakeValue = mock(Object.class);
-    map.save(fakeKey, fakeValue);
-
-    assertThat(map.typeToIdSize()).isEqualTo(1);
-    assertThat(map.enumToIdSize()).isEqualTo(1);
-    assertThat(map.pendingTypeToIdSize()).isEqualTo(0);
-    assertThat(map.pendingEnumToIdSize()).isEqualTo(0);
-  }
+public class PeerTypeRegistrationReverseMapTest extends TypeRegistryReverseMapTest {
 
   @Test
   public void saveToPendingCorrectlyAddsOnlyToPendingMaps() {
@@ -162,18 +130,6 @@ public class PeerTypeRegistrationReverseMapTest {
     assertThat(map.enumToIdSize()).isEqualTo(0);
     assertThat(map.pendingTypeToIdSize()).isEqualTo(0);
     assertThat(map.pendingEnumToIdSize()).isEqualTo(0);
-  }
-
-  private void addEnumInfoToMap(PeerTypeRegistrationReverseMap map) {
-    EnumId enumId = mock(EnumId.class);
-    EnumInfo enumInfo = mock(EnumInfo.class);
-    map.save(enumId, enumInfo);
-  }
-
-  private void addPdxTypeToMap(PeerTypeRegistrationReverseMap map) {
-    Integer pdxId = map.typeToIdSize();
-    PdxType pdxType = mock(PdxType.class);
-    map.save(pdxId, pdxType);
   }
 
   private void addEnumInfoToPendingMap(PeerTypeRegistrationReverseMap map) {
